+++ src/main/java/com/bitbank/services/UsersService.java
@@ -17,7 +17,7 @@ public class UsersService {
 	private UsersRepository usersRepository;
 
 	/**
-	 * Return single transaction.
+	 * Return single user.
 	 * 
 	 * @param id
 	 * @return Optional<UsersEntity>
@@ -40,4 +40,19 @@ public class UsersService {
 		}
 		return user.get().getId();
 	}
+
+	/**
+	 * Store the temp secret into user
+	 */
+	public boolean setTempSecret(Long id, String tempSecret) {
+		Optional<UsersEntity> usersEntity = usersRepository.findById(id);
+		if (usersEntity.isEmpty()) {
+			String error = "no user found with id " + id.toString();
+			throw new EntityNotFoundException(error);
+		}
+		usersEntity.get().setTempSecret(tempSecret);
+		UsersEntity updateUsersEntity = usersRepository.save(usersEntity.get());
+		return updateUsersEntity.getTempSecret().equals(usersEntity.get().getTempSecret());
+	}
+
 }