+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Next version
 + Change OTP library
 + Add EntityNotFoundException
++ Add method to store a temp secret
 
 ## 1.1.0
 + Upgrade dependencies